Vets::Model file/namespace reorg
