+++ lib/vets/model.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'vets/attributes'
+require 'vets/model/attributes'
 require 'vets/model/dirty'
 require 'vets/model/sortable'
 require 'vets/model/pagination'
@@ -16,7 +16,7 @@ module Vets
     extend ActiveSupport::Concern
     include ActiveModel::Model
     include ActiveModel::Serializers::JSON
-    include Vets::Attributes
+    include Vets::Model::Attributes
     include Vets::Model::Dirty
     include Vets::Model::Sortable
     include Vets::Model::Pagination
