+++ lib/vets/model/type/iso8601_time.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+require 'vets/model/type/base'
+
+module Vets
+  module Model
+    module Type
+      class ISO8601Time < Base
+        def self.primitive
+          ::String
+        end
+
+        def cast(value)
+          return nil if value.nil?
+
+          Time.iso8601(value)
+        rescue ArgumentError
+          raise TypeError, "#{@name} is not iso8601"
+        end
+      end
+    end
+  end
+end
