+++ lib/vets/model/type/titlecase_string.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+require 'vets/model/type/base'
+
+module Vets
+  module Model
+    module Type
+      class TitlecaseString < Base
+        def self.primitive
+          ::String
+        end
+
+        def cast(value)
+          return nil if value.nil?
+
+          value.to_s.downcase.titlecase
+        end
+      end
+    end
+  end
+end
