+++ lib/vets/model/type/utc_time.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+require 'vets/model/type/base'
+
+module Vets
+  module Model
+    module Type
+      class UTCTime < Base
+        def self.primitive
+          ::Time
+        end
+
+        def cast(value)
+          return nil if value.nil?
+
+          Time.parse(value.to_s).utc
+        rescue ArgumentError
+          raise TypeError, "#{@name} is not Time parseable"
+        end
+      end
+    end
+  end
+end
