--- lib/vets/type/utc_time.rb
@@ -1,21 +1,0 @@
-# frozen_string_literal: true
-
-require 'vets/type/base'
-
-module Vets
-  module Type
-    class UTCTime < Base
-      def self.primitive
-        ::Time
-      end
-
-      def cast(value)
-        return nil if value.nil?
-
-        Time.parse(value.to_s).utc
-      rescue ArgumentError
-        raise TypeError, "#{@name} is not Time parseable"
-      end
-    end
-  end
-end
